Refactor nodal analysis structure and add phase-aware OPR
