+++ src/Applications/NodalAnalysis/config/fieldConfigs.ts
