+++ src/Applications/NodalAnalysis/types.ts
@@ -1,0 +1,5 @@
+export type FrictionModel =
+  | 'Chen (1979)'
+  | 'Swamee-Jain'
+  | 'Colebrook-White'
+  | 'Laminar (auto)';
